refactor: extract CommunityStats component
